+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Projcet: 100 stories
+### Changed
+- Updated: storybuilder (v0.1.0)
 
 ## [0.0.2] - 2019-04-07
 ### Added